+++ Emuera/UI/Game/EmueraConsole.Print.cs
@@ -481,8 +481,7 @@ internal sealed partial class EmueraConsole : IDisposable
 	// public void PrintImg(string str)
 	public void PrintImg(string name, string nameb, string namem, MixedNum height, MixedNum width, MixedNum ypos)
 	{
-		//printBuffer.Append(new ConsoleImagePart(str, null, 0, 0, 0));
-		printBuffer.Append(new ConsoleImagePart(name, nameb, namem, height, width, ypos));
+		printBuffer.Append(new ConsoleImagePart(name, nameb, namem, height, width, ypos, null, DisplayMode.Relative));
 	}
 	#endregion